move middleware into own subpackage
